// source/frontend/src/models/api/generated/ApiGen_Concepts_ResearchFile.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { UtcIsoDate } from '@/models/api/UtcIsoDateTime';

import { ApiGen_Base_CodeType } from './ApiGen_Base_CodeType';
import { ApiGen_Concepts_File } from './ApiGen_Concepts_File';
import { ApiGen_Concepts_Organization } from './ApiGen_Concepts_Organization';
import { ApiGen_Concepts_Person } from './ApiGen_Concepts_Person';
import { ApiGen_Concepts_ResearchFileProject } from './ApiGen_Concepts_ResearchFileProject';
import { ApiGen_Concepts_ResearchFileProperty } from './ApiGen_Concepts_ResearchFileProperty';
import { ApiGen_Concepts_ResearchFilePurpose } from './ApiGen_Concepts_ResearchFilePurpose';

// LINK: @backend/apimodels/Models/Concepts/ResearchFile/ResearchFileModel.cs
export interface ApiGen_Concepts_ResearchFile extends ApiGen_Concepts_File {
  roadName: string | null;
  roadAlias: string | null;
  fileProperties: ApiGen_Concepts_ResearchFileProperty[] | null;
  requestDate: UtcIsoDate | null;
  requestDescription: string | null;
  requestSourceDescription: string | null;
  researchResult: string | null;
  researchCompletionDate: UtcIsoDate | null;
  isExpropriation: boolean | null;
  expropriationNotes: string | null;
  requestSourceType: ApiGen_Base_CodeType<string> | null;
  requestorPerson: ApiGen_Concepts_Person | null;
  requestorOrganization: ApiGen_Concepts_Organization | null;
  researchFilePurposes: ApiGen_Concepts_ResearchFilePurpose[] | null;
  researchFileProjects: ApiGen_Concepts_ResearchFileProject[] | null;
}